requirements.txt
Flask
gunicorn
Flask-SQLAlchemy
python-dotenv
psycopg2-binary